refactor: coordinates module
Collects functions for manipulation of atomic coordinates that used to be in lib.base
