+++ src/viperleed/calc/classes/slab/base_slab.py
@@ -34,11 +34,11 @@ from viperleed.calc.classes.atom_containers import AtomContainer, AtomList
 from viperleed.calc.classes.layer import Layer, SubLayer
 from viperleed.calc.classes.sitetype import Sitetype
 from viperleed.calc.lib import leedbase
-from viperleed.calc.lib.base import COLLAPSE_EPS
-from viperleed.calc.lib.base import add_edges_and_corners
-from viperleed.calc.lib.base import collapse
-from viperleed.calc.lib.base import collapse_fractional
 from viperleed.calc.lib.base import rotation_matrix_order
+from viperleed.calc.lib.coordinates import COLLAPSE_EPS
+from viperleed.calc.lib.coordinates import add_edges_and_corners
+from viperleed.calc.lib.coordinates import collapse
+from viperleed.calc.lib.coordinates import collapse_fractional
 from viperleed.calc.lib.itertools_utils import pairwise
 from viperleed.calc.lib.string_utils import parent_name
